requirements.txt
pandas>=2.1.0
PyQt6>=6.6.0
matplotlib>=3.8.0
pillow>=10.1.0 